Move commands to one file
